Object.Types

Documentation

type family Output object action

class Action object action where

Methods

(.) :: object -> action -> Output object action infixl 8

data a := b infixl 9

Constructors

a := b infixl 9

type family MethodOutput object method

(%) :: b -> (b -> c) -> c infixl 0

(...) :: (t1 -> t) -> (t2 -> t1) -> t2 -> t

(*>) :: Functor f => f a -> (a -> b) -> f b infixl 4

Explanation of all these operators. They are basically an assortment of Functor and Applicative functions. I.e., '[1,2,3] *> (+1)' === map (+1) [1,2,3]' The asterisk * indicates where the Functor, Applicative is. And the greather than (smaller than) indicates in which direction the argument is going. So you could also have '[(+1),(+2),(+3)] *< 1 === map ($ 1) [(+1),(+2),(+3)]'. If there are asterisks on both sides the Functor has to be Applicative

(*<) :: Functor f => f (a -> b) -> a -> f b infixr 4

(*.) :: (Action a action, Functor f) => f a -> action -> f (Output a action) infixl 4

(>*) :: Functor f => b1 -> f (b1 -> b) -> f b infixl 4

(<*) :: Functor f => (a -> b) -> f a -> f b infixr 4

(.*) :: (Action object a, Functor f) => object -> f a -> f (Output object a) infixl 4

(*>*) :: Applicative f => f a -> f (a -> b) -> f b infixl 4

(*<*) :: Applicative f => f (a -> b) -> f a -> f b infixr 4

(*.*) :: (Action a1 a, Applicative f) => f a1 -> f a -> f (Output a1 a) infixl 4